backend/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
ultralytics
opencv-python
tensorflow
firebase-admin
cloudinary
